removed rotation
So, this is a bit more difficult. Most blocks in mc have a locked rotation (so planks are always the same). The framed drawer does not, so if you were to use planks, only 1 way it would look right. I've "fixed" this by using uvlock, hover, now blocks like a furnace (that have rotations) only work in 1 direction... I think this is a choice of the "lesser" of 2 evils.